requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

# Core dependencies
python-dotenv>=1.0.0
# Core LangChain packages
langchain==0.3.8
langchain-core==0.3.22
langchain-community==0.3.7
langchain-openai==0.2.9
langchain-text-splitters==0.3.1
transformers==4.57.1
sentencepiece>=0.1.99
# Optional GPU version (CUDA 12.2 example)
torch==2.9.1+cu128
torchvision==0.15.2+cu128
torchaudio==2.1.0

# API clients
openai>=1.0.0
google-search-results>=2.4.2
stravalib
geopy

# CLI and UI dependencies
rich>=13.0.0
prompt-toolkit>=3.0.0

# Local/Open-source model support
langchain-ollama>=0.1.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3 # For web scraping

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
ruff>=0.1.0
mypy>=1.0.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0


